refactor(aoscc): move AosccDetail out from components folder
